Don't fully initialize uACPI unless absolutely needed
Main concern for doing that is that running AML might put the
hardware into a non-PC compatible state with no reliable way
to reverse that (as per comment on #54).

For now `acpi_init` only sets up "early table access", as it
previously did, to allow ACPI table enumeration, but if the need
arises one can call `acpi_full_init` to be able to call into
uACPI functions that require the namespace to be loaded and initialized.